+++ certlogic/certlogic-js/src/test/validation/test-data-accesses.ts
@@ -1,0 +1,22 @@
+const { deepEqual } = require("chai").assert
+
+import { CertLogicOperation } from "../../typings"
+import { dataAccessesWithContext } from "../../validation"
+
+
+describe("dataAccessesWithContext", () => {
+
+    it("should produce correct context (1)", () => {
+        const expr = { "var": "ci" }
+        deepEqual(dataAccessesWithContext(expr), { "ci": [ expr ] })
+    })
+
+    it("should produce correct context (2)", () => {
+        const expr: CertLogicOperation = { "extractFromUVCI": [ { "var": "ci" }, 1 ] }
+        deepEqual(
+            dataAccessesWithContext(expr),
+            { "ci": [ expr ] })
+    })
+
+})
+